fix(cli): skip master password prompt when company has no encryption
- Move isEncrypted and VaultProtocolState to shared package for reuse
- Update CLI login to only store master password if company has encryption
- Clear stored password on login if company doesn't have encryption
- Remove prompt for new password in requireMasterPassword (throw instead)
- Refactor web vaultProtocol.ts to re-export from shared

This fixes the issue where CLI would hang waiting for master password
input during non-interactive scripts like ./go system up, even when
the company hasn't enabled vault encryption.
